// src/app/core/service/model/praise-discipline-type.model.ts
export interface PraiseDisciplineTypeModel {
  id?: string;
  creationTime?: string;
  content?: string;
  duocCongDiemHanhKiem?: boolean;
  plusMarkConduct?: number;
  duocCongDiemTotNghiep?: boolean;
  plusMarkGraduation?: number;
  biTruDiemHanhKiem?: boolean;
  minusMarkConduct?: number;
  type?: number;
}